first 2 pages up and running
