+++ public/stylesheets/styles.css
@@ -1,0 +1,58 @@
+header h1 {
+  text-align: center;
+  font-weight: bold;
+  font-size: 32px;
+  border: 0px; }
+header .top-header-text p {
+  text-align: center;
+  font-size: 30px;
+  margin: 0px; }
+header .pester-me p {
+  text-align: center;
+  font-size: 12px;
+  color: gray;
+  margin: 10px 0px 10px 0px; }
+
+article .input-box {
+  width: 400px;
+  height: 170px;
+  background-color: white;
+  margin: 10px calc(50% - 200px);
+  border: solid gray 1px; }
+  article .input-box h2 {
+    text-align: center;
+    font-weight: bold; }
+  article .input-box input {
+    height: 30px;
+    width: 200px;
+    margin: 0px auto;
+    text-align: center; }
+  article .input-box button {
+    margin: 20px auto;
+    margin-left: auto;
+    background-color: yellow;
+    border: solid gray 1px; }
+
+article h3 {
+  font-size: 20px;
+  color: gray;
+  text-align: center; }
+
+.legislator {
+  clear: left;
+  padding: 10px; }
+  .legislator .legislator-info {
+    float: left; }
+    .legislator .legislator-info h4 {
+      font-size: 24px;
+      font-weight: bold;
+      float: left;
+      padding: 10px 0 5px 10px; }
+    .legislator .legislator-info h5 {
+      font-size: 16px;
+      clear: left;
+      padding: 5px 0 0 10px; }
+  .legislator img {
+    float: left; }
+
+/*# sourceMappingURL=styles.css.map */
