+++ public/stylesheets/styles.scss
@@ -1,0 +1,85 @@
+header{
+  h1 {
+    text-align: center;
+    font-weight: bold;
+    font-size: 32px;
+    border: 0px;
+  }
+
+  .top-header-text {
+    p {
+      text-align: center;
+      font-size: 30px;
+      margin: 0px;
+    }
+  }
+
+  .pester-me {
+    p {
+      text-align: center;
+      font-size: 12px;
+      color: gray;
+      margin: 10px 0px 10px 0px;
+    }
+  }
+}
+
+article .input-box {
+  width: 400px;
+  height: 170px;
+  background-color: white;
+  margin: 10px calc(50% - 200px);
+  border: solid gray 1px;
+
+  h2 {
+    text-align: center;
+    font-weight: bold;
+  }
+
+  input {
+    height: 30px;
+    width: 200px;
+    margin: 0px auto;
+    text-align: center;
+  }
+
+  button {
+    margin: 20px auto;
+    margin-left: auto;
+    background-color: yellow;
+    border: solid gray 1px;
+  }
+}
+
+article{
+  h3 {
+    font-size: 20px;
+    color: gray;
+    text-align: center;
+  }
+}
+
+.legislator {
+  clear: left;
+  padding: 10px;
+
+  .legislator-info{
+    float: left;
+
+    h4 {
+      font-size: 24px;
+      font-weight: bold;
+      float: left;
+      padding: 10px 0 5px 10px;
+    }
+
+    h5 {
+      font-size: 16px;
+      clear: left;
+      padding: 5px 0 0 10px;
+    }
+  }
+  img {
+    float: left;
+  }
+}
