+++ views/html/homepage.html
@@ -1,0 +1,46 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>Find Legislator Homepage</title>
+  <meta charset="UTF-8">
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
+  <link rel="stylesheet" href="styles.css">
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
+  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
+</head>
+
+<body>
+  <div class="jumbotron">
+    <header class="container">
+      <div class="row">
+        <div class="col-lg-12">
+          <h1>What Have You Done?</h1>
+        </div>
+      </div>
+      <div class="row">
+        <div class="col-lg-12 top-header-text">
+          <p>Discover the <strong>voting records</strong></p>
+          <p>of your <strong>local representatives</strong></p>
+        </div>
+      </div>
+      <div class="row">
+        <div class="col-lg-12 pester-me">
+          <p>Then pester their offices until their views reflect yours!</p>
+        </div>
+      </div>
+    </header>
+    <article class="container">
+      <div class="row">
+        <div class="col-lg-12 input-box">
+          <h2>Enter Your Zip Code</h2>
+          <input class="form-control" type="text" />
+          <div class="text-center">
+            <button type="button" class="btn btn-md">Go!</button>
+          </div>
+        </div>
+
+      </div>
+    </article>
+  </div>
+</body>
+</html>
