+++ views/html/zipcodePage.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>Zipcode Page</title>
+  <meta charset="UTF-8">
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
+  <link rel="stylesheet" href="styles.css">
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
+  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
+</head>
+
+<body>
+  <div class="jumbotron">
+    <header class="container">
+      <div class="row">
+        <div class="col-lg-12">
+          <h1>What Have You Done?</h1>
+        </div>
+      </div>
+    </header>
+    <article class="container">
+      <div class="row">
+        <div class="col-lg-12 house">
+          <h3>House</h3>
+        </div>
+        <div class="col-lg-12 senate">
+          <h3>Senate</h3>
+        </div>
+
+      </div>
+    </article>
+  </div>
+</body>
+</html>
