+++ views/partials/homepage.handlebars
@@ -1,0 +1,33 @@
+<div class="jumbotron">
+  <header class="container">
+    <div class="row">
+      <div class="col-lg-12">
+        <h1>What Have You Done?</h1>
+      </div>
+    </div>
+    <div class="row">
+      <div class="col-lg-12 top-header-text">
+        <p>Discover the <strong>voting records</strong></p>
+        <p>of your <strong>local representatives</strong></p>
+      </div>
+    </div>
+    <div class="row">
+      <div class="col-lg-12 pester-me">
+        <p>Then pester their offices until their views reflect yours!</p>
+      </div>
+    </div>
+  </header>
+  <article class="container">
+    <div class="row">
+      <div class="col-lg-12 input-box">
+        <h2>Enter Your Zip Code</h2>
+        <div class="text-center">
+          <form action="./zipcodepage" method="get">
+            <input class="form-control" type="text" name="zipcode" />
+            <!--<button type="button" class="btn btn-md">Go!</button>-->
+            <input type="submit" value="submit" />
+          </form>
+        </div>
+      </div>
+    </div>
+  </article>
